Moved rule query building logic to separate query_builders.py file
